requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf_spectacular==0.27.2
groq==0.11.0
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.0
